feat(#55): Add backend for room temperature control
