+++ smart-home-simulator/src/main/resources/static/js/dashboard.js
@@ -1,4 +1,8 @@
 var dashboardContext;
+var shhTab;
+var shhRoom;
+var shhZone;
+
 window.onload = async function () {
     const response = await fetch("/dashboard", {method: "GET"});
     let responseData = await response.json();
@@ -6,7 +10,9 @@ window.onload = async function () {
     dashboardContext = new Vue({
         el: "#dashboardContextContent",
         data: {date: responseData.date, time: responseData.time, layout: responseData.fileName, tempOut: responseData.tempOut, location: 'Placeholder'}
-    })
+    });
+
+    loadSHHTab();
 }
 
 function openModule(evt, modName) {
@@ -152,4 +158,79 @@ async function offLights(e, room){
     console.log(responseData);
 }
 
+async function loadSHHTab(){
+    const getSHH = await fetch("/dashboard/shh", {method: "GET"});
+    let shhTabResponseData = await getSHH.json();
 
+    const getRooms = await fetch("/dashboard/shhRooms", {method: "GET"});
+    let shhRoomResponseData = await getRooms.json();
+
+    shhTab = new Vue({
+        el: "#shhVariables",
+        data: {selected: "null", zones: shhTabResponseData, selectedZone:{}},
+        methods:{
+            onSelected(event) {
+                let selectedIndex = event.target.value
+                this.selectedZone =  this.zones[selectedIndex]
+            },
+            async changeZone(e){
+                e.preventDefault();
+                const data = JSON.stringify(this.selectedZone);
+                const response = await fetch("/dashboard/shhChangeZone", {method: "POST", body: data, headers: {
+                        "Content-Type": "application/json",
+                    }});
+                let responseData = await response.json();
+                this.zones = responseData;
+                return;
+            }
+        }
+    });
+    shhRoom = new Vue({
+        el: '#shhRoomTemperatures',
+        data: {
+            rooms: shhRoomResponseData,
+            zones: shhTabResponseData,
+            roomSelectedZone: "-1",
+        },
+        methods:{
+            async onRoomSelected(event, roomName) {
+                let object = {name: roomName, zoneID: event.target.value}
+                const data = JSON.stringify(object);
+                const response = await fetch("/dashboard/shhRoomChangeZone", {method: "POST", body: data, headers: {
+                        "Content-Type": "application/json",
+                    }});
+            },
+            async overrideTemperature(roomName, event){
+                let object = {name: roomName, temp: event.target.value}
+                const data = JSON.stringify(object);
+                const response = await fetch("/dashboard/shhOverrideRoomTemperature", {method: "POST", body: data, headers: {
+                        "Content-Type": "application/json",
+                    }});
+                console.log(response);
+            },
+        }
+    });
+
+    shhZone = new Vue({
+        el: '#shhZone',
+        data:{
+            name: '',
+            setting: false,
+            temperature: 0
+        }
+    });
+}
+
+async function addZone(e){
+    e.preventDefault();
+    let object = {name: shhZone.name, setting: shhZone.setting, temperature: shhZone.temperature};
+
+    let data = JSON.stringify(object);
+
+    const response = await fetch("/dashboard/shhAddZone", {method: "POST", body: data, headers: {
+            "Content-Type": "application/json",
+        }});
+    let responseData = await response.json();
+    shhTab.zones = responseData;
+    shhRoom.zones = responseData;
+}
